+++ RunAvgWeight.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: RunAvgWeight.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 //    DATE: 2024-06-30
 // PURPOSE: Arduino library to calculate the running average with weights by means of a circular buffer
 //     URL: https://github.com/RobTillaart/RunAvgWeight
@@ -31,7 +31,7 @@ RunAvgWeight::~RunAvgWeight()
 
 
 //  resets all counters
-void RunAvgWeight::clear()
+bool RunAvgWeight::clear()
 {
   _count = 0;
   _index = 0;
@@ -39,21 +39,26 @@ void RunAvgWeight::clear()
   _sumWeights = 0.0;
   _min = NAN;
   _max = NAN;
+  if (_array == NULL)
+  {
+    return false;
+  }
   for (uint16_t i = _size; i > 0; )
   {
     // setting values and weight to zero keeps addValue simpler
     _values[--i] = 0.0;
     _weights[--i] = 0.0;
   }
+  return true;
 }
 
 
 //  adds a new value to the data-set
-void RunAvgWeight::addValue(const float value, const float weight)
+bool RunAvgWeight::addValue(const float value, const float weight)
 {
     if ((_values == NULL) || (_weights == NULL))
     {
-      return;
+      return false;
     }
 
   _sumValues -= _values[_index];
@@ -74,6 +79,7 @@ void RunAvgWeight::addValue(const float value, const float weight)
 
   //  update count as last otherwise if ( _count == 0) above will fail
   if (_count < _size) _count++;
+  return true;
 }
 
 
@@ -190,6 +196,20 @@ float RunAvgWeight::getStandardError()
   temp = temp/sqrt(n);
 
   return temp;
+}
+
+
+//  Return coefficient of variation.
+//  If buffer is empty or has only one element or zero average, return NAN.
+float RunAvgWeight::getCoefficientOfVariation() const
+{
+  float temp = getStandardDeviation();
+  if (temp == NAN) return NAN;
+  if (_sum == 0) return NAN;
+
+  //  float cv = temp * getFastAverage();
+  float cv = temp * _sumWeights / _sumValues;
+  return cv;
 }